webdriverio: refactor find-strategy method (#3778)
## Proposed changes
Main objective was to split conditions with further logic for assignment strategy and value. In my opinion findStrategy function becomes clearer and more understandable. After refactoring we have:

- one "private" (help function that is not export) defineStrategy. Here we put all conditions and don't think about anymore else
- the same public method findStrategy. Here we think about what using and value should be for each case. We don't bother by what should be for this conditions.

PREVIOUS COMMENTS in PR #3741 

[//]: # (Describe the big picture of your changes here to communicate to the maintainers why we should accept this pull request. If it fixes a bug or resolves a feature request, be sure to link to that issue.)

## Types of changes

[//]: # (What types of changes does your code introduce to WebdriverIO?)
[//]: # (_Put an `x` in the boxes that apply_)

- [x] Refactoring (non-breaking change)

## Checklist

[//]: # (_Put an `x` in the boxes that apply. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your code._)

- [x] I have read the [CONTRIBUTING](https://github.com/webdriverio/webdriverio/blob/master/CONTRIBUTING.md) doc

## Further comments

[//]: # (If this is a relatively large or complex change, kick off the discussion by explaining why you chose the solution you did and what alternatives you considered, etc...)

### Reviewers: @webdriverio/technical-committee
